pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dataflow-cv"
version = "0.1.0"
description = "A data processing library for computer vision datasets"
readme = "README.md"
authors = [
    {name = "DataFlow Team"}
]
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]
keywords = ["computer-vision", "dataset", "conversion", "visualization", "coco", "yolo", "labelme"]
dependencies = [
    "numpy>=1.19.0",
    "opencv-python>=4.5.0",
    "Pillow>=8.0.0",
    "click>=8.0.0",
]
requires-python = ">=3.8"

[project.optional-dependencies]
full = [
    "pycocotools>=2.0.0",
    "torch>=1.9.0",
    "torchvision>=0.10.0",
]

[project.scripts]
dataflow = "dataflow.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["dataflow*"]
exclude = ["tests*", "examples*"]